build: use multiple crates
The project was growing in size, and I believed that embedding the
database logic into the same crate that defines the CLInvoice schema
would be a bad idea.

After all, the data does not completely depend on MongoDB. We could just
as easily adapt this scheme to Oracle or PostreSQL if that were desired
in the future. Should that be desired, it would be incredibly
complicated to create new bindings for the concrete implementation of
the data structures unless they were separated fundamentally— which is
what this commit does.

Additionally, this will allow freedom of exploration for the
command-line interface, which can now be worked on independently of
other modules.