ref(adapters): use TOML for adapter testing
Initially I was trying to use Postgres to get this project off the
ground (actually, I was trying to use Mongo, _then_ I switched to
Postgres…) but now I've decided that I should cut back on the amount of
overhead in order to ensure that the design of the application is up to
snuff.

Once I get the kinks ironed out (including CLI commands) I can implement
a second adapter, for Postgres.

I want to try to come up with a better way to present implementation
than to force maintainers to copy all of the files under
`clinvoice_adapter_toml/data` into a new directory just so they can run
some regex replace commands on the names. I think there must be a better
way, since that is a lot of code duplication (even for wrapper classes).