style: closer imports come first
I used to have more distant/foreign imports be placed first in the text
files, but now I'm using them at the very top to signify that they are
the closest to the current file.

Sure, this oculd have been discerned just from looking at the import
path, but having a consistent structure is also useful in immediately
identifiying which things might have to be altered when refactoring.