fix: use Option<> where necessary; remove Set<>s
Before, we did not use Option<> because it was implicit on `AnyValue`.
With `RetrieveWhen`, it is explicit and we must specify when the values
could be `None`.

We also used `HashSet`s in order to describe a list of acceptable
values, but this was ambiguous. Instead, `RetrieveWhen` allows for sets
of a datatype to be passed so that matching operations can be performed.