test(clinvoice_adapter_bincode): bench read+deser
I saw create() was a potentially expensive method, so I wanted to know
how much of the time in that test was spent just reading and
deserializing files.

Knowing this will allow me to choose how to implement retrieve, since my
initial plan was to use `rg`; however using Bincode will make the files
look nothing like the Displayed structs so this is not possible.