feat(util): read_files
This method ensures that all paths passed back to the caller are valid
files.

Somehow, invalid files are still being received. This must mean that
files are being destroyed and/or recreated inbetween the method being
iterated over and a check being performed. Honestly I think this is just
a filesystem issue and it can't be reproduced consistently so I'm not
worried about it.