ref(clinvoice_adapter): borrow Store for adapters
This is so that we can make the XYZAdapter types (not traits) contain a
borrowed Store. The reason for this is that we want to be able to return
Store as an owned type, and we don't need to bring the types out of
scope (only what they adapt, ergo XYZ).

We're going to move to a model of wrapping the adapted type into the
adapter with a borrow of the store in order to perform operations so
that returning and passing complicated types is de-incentivized. We can
always derive the correct adapter from the store anyway.