fix(clinvoice_adapter_bincode): avoid `Self` where necessary
This change is related to a Reddit thread (linked below).

Using `Self` created an E0505 and E0597 combination. I'm not fully sure
why— my best guess is that `Self` was bound to `<'_, 'store>` but
qualifying `BincodeXYZ::<'_, 'store>` also fixes this issue. For some
reason, just writing out the types here is enough for Rust to understand
what we're trying to do.

This would have stumped be but thanks to the Rust community on Reddit, I
was able to keep going.

RELATED: r/rust/comments/lyiiwg/misunderstanding_e0505/