fix(io input util): add lifetime args to DynResult
This will resolve the troublesome `L::retrieve` statement from
`retrieve_locations_or_err`. This was the original reason we started to
refactor the adapters anyway. The reason we couldn't just add this to
begin with was that the adapters had _three_ lifetimes attached— the
refactored adapters have just one (which resolves this issue).