feat(MatchWhen): derive De/Serialize
This required changing our raw borrows to `Cow`s, due to restrictions on
`#[serde(borrow)]`. More specifically, we had to specify:

```rust
serde(bound(deserialize = "&'element T : Deserialize<'de>"))
```

But this would almost never be fulfilled. For example, deserializing a
`MatchWhen<usize>` would fail because `&usize : Deserialize<'de>` is not
implemented. Therefore, we had to change our borrowed types to some
other owned type. However, we don't want to actually own the data,
because we only ever want to borrow it in order to run the queries (in
order to prevent unecessary duplication every query, especially in cases
where multiple complicated queries are appended together). Overall this
is just undersirable.

`Cow` seems like the perfect fit for this case, as it will borrow types
until the first time they are written to. Since we never need to write
to the types outside of deserializing, we can transparently borrow but
allow a clone + write if necessary. This preserves performance in most
cases but gives permissiveness necessary for deriving De/Serialize.