feat(clinvoice_data export): initial impl of Markdown
The idea behind this approach (in comparison to the original CLInvoice)
is that we want to provide tools for each entity to implement the
`IntoMarkdown` trait that will consistently render as valid markdown.
Before, we were just writing it by hand and there was always a chance
that the markdown could have one tiny error in a small place and it
would be difficult to find it.