feat(grammars): add initial query grammars
We needed something to show users what the syntax of the queries are.
Rust users should be fairly used to ungrammars, and ungrammars are
fairly easy to read for those who are not familiar with them.

I intend to add a new `clinvoice_grammar` crate that can resolve the
'//#include`s at the top of each file. The idea of those is that
sometimes rules from other files need to be included in the output, but
they should be extracted from each other because they are referenced
multiple times. A custom parser will be written to identify these and
add them to the output where necessary.