ref(grammars): optimize imports
This commit moves certain Rules around to the places where they are
references, and organizes the order of rules so that complex rules are
introduced first (and only then, the lesser rules which they are made
of). This should make the output of linking files more consistent and
more natural to read.