chore: remove unneeded `'store` lifetime from Adapters
This is a relic from a refactor several weeks ago. This lifetime is no
longer needed, as the correct lifetimes of all the `Store` borrows can
simply be inferred on a per-call basis rather than insisting all borrows
are from the same lifetime in order to function with the trait.

This saves from having to write `<'store>` every time we want to write a
generic function for an adapter.