ref(Create): generic `create_` functions
This will allow other adapters to easily _adapt_ functionality for the
`Create` command.

The reason for using macros on `Retrieve` and not `Create`, was that the
operations used in `Retrieve` are more complicated are require more
generic arguments. Using a macro for `Retrieve` actually resorted in a
more simple, less-SLOC solution to the problem.

The issue with macros, however, is that they lack autocomplete for their
variables while inside the scope (and how could they provide it, given
anything can be an `ident` or `expr`). Functions are preferred for this
reason, but it was not so easily doable in `Retrieve`.