feat(clinvoice_export markdown): impl `Display`
I was using a custom `render` function which was converting each type to
a string, but I came to realize that implementing `Display` would be a
better option. I believe this is so, because when using `format!()` I
would not have to convert `Element`s or `Text`s to `String` manually—
instead, they will be written to the `Formatter`.