ref(clinvoice_adapter data): reword `Error::NoData`
The reason for this, is that `NoData` is a category of `Error`s that
should be usable in more contexts than just when setting up `clinvoice`
for the first time. It should also apply to when a query has failed to
retrieve any data (when some data was needed, such as for
`clinvoice retrieve employee --set-default`).