feat: special `Match` for strings
In testing this application, it became abundantly clear that the `Match`
type does not work for strings. Strings require a very different
approach to matching, one that allows checking for "contains" and
regular expressions rather than "HasAny" or "InRange". While `Match`
remains a reliable type for generic data, strings are the special case
here.

This commit adds the `MatchStr` type for use with any type which
implements `AsRef<str>`. It will allow users to perform more
comprehensive searching on string data.