feat: allow users to control exporting contact info
Before, contact information had to be exported either in total or not at
all. This is problematic, because `Employee`s may have personal pieces
of information (although however discouraged this may be) as part of
this system.

This commit allows users to control what information is distributed
publicly with `Job`s, and which information is just tracked for internal
use.