fix(input): empty `select_one` crash w/out helpful err
If users attempted to update `Job`s, for instance, without creating
any, it would crash `clinvoice` without telling the user what to do to
fix the `Error`. It would simply say "empty list of items passed to
select". To fix this, we return a `clinvoice_adapter::data::Error::NoData`
in this case (which tells the user that they need to add data if they
wish to perform the operation they attempted).