fix(clinvoice_query): recursive evaluation of bounds
A frequent issue with this module was that the compiler would recurse >
128 times trying to evaluate the bounds of arguments. This commit turns
`mut arg: impl Iterator` into `arg: &mut impl Iterator` so that no
evaluation of bounds on `mut` is necessary— only `impl Iterator`. This
change allows `&mut` to be specifically passed as part of the function
call, which relieves the troubles that the compiler was having.