ref(clinvoice_query): combine `Invoice` and `InvoiceQuery`
With the previous model, users were unable to get _all_ `Invoice`s in
the system. They either had to get `Invoice`s which had not been issued,
or an `Invoice` which had been issued. This new model allows users to
query both or query any.

It does not, however, prevent users from querying `Invoice`s which have
been paid but not issued. This is an impossible state, however, because
of the underlying non-view model of the data. `Invoice`s must be sent
before being paid.