feat(Match): AllGreaterThan, AllLessThan, AllInRange
Certain implementations of the operations for `Match`ing had
`.iter().any()` in their bodies. These implementations did not support
operations which asserted all of a set met a specific critera— when this
may be a desired function. As such, these three conditions have been
added with `all()` variants to ensure all of the elements match.

NOTE: `HasAll` exists as `Not(HasAny())`