fix(Time): `stop` only when a `Timesheet` is active
Before, when attempting to `input::select_one` a `Timesheet` to stop
working on, there was no check to make sure that there was actually an
active `Timesheet` to stop. Now we check first, and if necessary error
to tell the user that there is no active `Timesheet` on their currently
selected `Job`.