ref(input util contact): query `Location` for Address
Before there was a single query done in order to select which addresses
would be added to the user's contact information. Now, a query is done
every time an address is added to the contact information.

This is more clear for the user, as they may not wish to add any contact
information regarding addresses, and they also may wish to have
completely different queries for their addresses.