style(clinvoice_adapter): allow to/into methods
These methods usually take an explicit `self` in the function signature.
However, this would not work for our adapters as we have no knowledge of
the structures which we are implementing. While normally this would
warrant a different naming scheme, we actually _do_ adhere to the
`&self`/`self` split of `to`/`into` (because the underlying types of
each adapter are ensured to be compatible with the naming) yet it is not
detected (and how could it detect this…)