feat(clinvoice_finance): draft `ExchangeRates` type
The `ExchangeRates` type will be used as an abstraction around scraping
the latest exchange rates from the ECB and checking periodically whether
or not the latest `ExchangeRates` are stored on disk.

`ExchangeRates` will be stored in the temporary directory, as it seems
unecessary to persist them longer than one session. In many cases,
they may not be used more than one time. If the struct detects that
there are no locally-stored exchange rates, it will download them and
load them. Similarly for outdated exchange rates.

If it detects that there are valid exchange rates (e.g. performing
multiple exchanges in rapid succession) it will simply be loaded from
disk.