ref(clinvoice_finance): use crate-level `Error`
Despite the fact that I personally dislike this pattern and find it to
be undescriptive (since functions no longer show which `Error`s they may
return, and instead show the entire domain of `Error`s that the entire
module may return), the lack of checked exceptions in Rust makes it
harder to do this. Using crate-level `Error`s is much more convenient.