ref(Money): pass in `ExchangeRates` for `exchange`
When performing an `exchange` before, we generated the `ExchangeRates`
internally. This is problematic for one reason: if we want to perform
many `exchange`s in quick succession, the `ExchangeRates` will have to
be read from disk and parsed over and over again. While this is not a
_long_ process, it will add time to the otherwise quick operation of
exchanging currencies.

Now, users have the option of maintaining their single instance of
`ExchangeRates` while performing looping operations (see `Job::total`
for an example). This will mean we only have to read from disk a single
time, and parse a single time as well.