perf(clinvoice_adapter_bincode): check `id` on retrieve
This performance improvement is based on the fact that there is one
aspect of an entity which can be read without deserializing: its `id`.
If a user specifies a `Match` on `Id`s, it will scan the filename to
make sure the filename `Match`es the `Id` before deserializing to
inspect further.