ref: remove `clinvoice_adapter_bincode`
In order to create better support for databases (which is how I envision
that most people will use this application, and the _only_ way that this
application can be used scalably…) code relating to filesystems must be
gutted. Active connections shouldn't be maintained while also only
communicating with the local operating system— trying to do both would
make the API awkward for either use case.

BREAKING CHANGE: this removes support for the Bincode adapter!