fix(adapters): reset to `BeforeClause` from `AfterClause`
This was to fix a case where if a `Match::Any` was written _after_ the
start of a `WHERE` clause, it would result in my implementations
generating another `WHERE` keyword. This is not valid SQL, so to fix it
I made `Any` return the same context as was passed in.