ref: don't use recursive self-join for LocationView query
This method of generating the query was problematic. It needed that you
be before the `WHERE` clause, but what happened if there were multiple
`Location`s being queried?

The second issue was that what if someone wants to look for all of the
organizations in the "USA". Previously, there was no way to do this. But
now, since it is in a recursive query, this is possible.