perf(clinvoice_finance): make async
CLInvoice needed a way to ensure that comparisons of `Money` amounts
inside the database were standardized, since there is no way to know
what currency some row is in simply by its name (e.g. `E.hourly_rate`).

Thus, all data is now recorded in EUR in the database, but users may
interact in whatever currency they are comfortable with and it will be
seamlessly exchanged whenever necessary.

Because this functionality is required where `async` is possible,
`reqwest` should not be used with `blocking` because it may cause the
current work being done to halt. As such, the entire `clinvoice_finance`
crate has been redesigned to work with an asynchronous runtime.