ref: merge `View` variants
This commit removes the difference between the `clinvoice_schema` and
`clinvoice_schema::views` types. Only the `views` were actually used,
and in practice it took more effort to convert from one to the other.

Now all we have are the views, which makes things simpler to reason
about. We don't care about how to store things in transit, only how they
should appear in memory.