ref(clinvoice_schema): `Vec<Contact>` for `contact_info`
This is desirable because the database is the one who is keeping track
of the label being unique for us. Not only that, but the `HashMap` is
probably overkill considering how I do not foresee the length of contact
information being stored go above maybe 20, which can definitely be
searched using iterative methods.

To do this, we move the key of the old `HashMap` into `contact.label`.