ref: use `Transaction` for `Updatable`
This is so that nothing in the database will be changed unless a series
of updates is successful.

This was not necessary for `delete` because it is not desirable to
cascade deletions to other tables except in specific circumstances, and
in those circumstances `ON DELETE CASCADE` has already been specified in
the database.