perf(PgLocation): inscrease performance of `PgLocation::update`
There are several techniques used to do this. Firstly, we use a
`LinkedList` to store the `Vec`s of `Location`s since we don't know how
many `Vec`s there will be. Second, we don't collect the `outer`s into a
`Vec` in the first place until we know that therer are any.