test(PgOrganization): begin implementing test
I started writing this test case a few days ago. I discovered that when
updating a given Organization, you would have to `DELETE FROM` the
`contact_information` table if it was not present when updating a given
`organization.contact_info`. The SQL looked like this:

```sql
DELETE FROM contact_information C WHERE
  -- These are all the primary keys for the contacts that were
     present when updating
  (C.organization_id = 1 AND NOT (label = 'foo' OR label = 'bar')) OR
  …
  (C.organization_id = 2 AND NOT (label = 'something else' OR label = 'another thing'))
;
```

However, the `clinvoice_match` crate was not capable of expressing this.
That led to the introduction of `MatchRow`, and then the simplification
of `clinvoice_schema`, etc.

TODO: fix errors resulting from changes to `clinvoice_schema`