ref(adapters): avoid writing unchecked SQL keywords
`sqlx::query!` checks syntax on compilation, but it is possible when
generating dynamic queries to misspell a keyword (e.g. `SELCT`). This
will be caught by the unit tests, but why not make such an error
impossible altogether?

The `QueryBuilderExt` now provides methods to write SQL keywords to a
`QueryBuilder` when it is necessary to do so raw.